video call button
These !importants are actually necessary. 

First, its required for .colorable-1bkp8v.white-3GPOIp .active-1QRrIS to be transparent as we change the color when we hover the button to hover-color (which, when different from main color, makes the svg's background different https://prnt.sc/17paaeu

Second, we need to add an important to background-color because apparently discord overwrites this with white. To avoid adding a second important to :hover, I use :not(:hover)

A rare usage lol